fix: window refresh problem on Windows from e95a5f9
This is a fix of previous fix e95a5f94aa7d27284ce00ee75cb067dd95e8fe7d.
It had solved a refresh issue on Linux when the VerticalScrolledFrame
interior frame size was changed. With the fix for Linux, it produces the same
issue on Windows. This commit fix the issues.