fix: wrong slice points with non zero start point
Slices point addresses of the electribe are relative to the start point.
The editor was addressing slices from the beginning of the sample,
not from the start point...
This is corrected here.
It has been verified that negative values of slice.start point are accepted
and correctly handled by the electribe (OS 2.02).